+++ release_notes.md
@@ -29,4 +29,5 @@
 - cloning of objects instead of simple copy in Specification
 - Constraints params now correctly set by addConstraint when the config is a String
 - from_dict can receive a string or an object. If the object is a Buffer (the result of a http request) it directly convert to JSON. This is a very common case.
+- result and ResultValues correctly adapted in rows and column format of the RI (from_dict and to_dict)
 - BUG scripting!